Renamed js vars/css names
