+++ app/assets/stylesheets/ltags/LinkTag.scss
@@ -167,7 +167,8 @@
   }
 }
 
-.crayons-article-sticky, .crayons-story__indention-display-ad {
+.crayons-article-sticky,
+.crayons-story__indention-billboard {
   .ltag__link {
     .ltag__link__org__pic {
       width: 60px;
