+++ app/views/admin/billboards/_form.html.erb
@@ -28,7 +28,7 @@
     </div>
 
     <div class="crayons-field">
-      <div id="display-ad-targeted-tags"></div>
+      <div id="billboard-targeted-tags"></div>
     </div>
 
     <div class="crayons-field hidden">
@@ -115,7 +115,7 @@
     </div>
   </div>
   <div>
-    <div class="crayons-card crayons-card--secondary crayons-sponsorship text-styles text-styles--display-ad">
+    <div class="crayons-card crayons-card--secondary crayons-sponsorship text-styles text-styles--billboard">
       <% if @billboard.persisted? %>
         <%= @billboard.processed_html.html_safe %>
       <% else %>
