+++ app/views/articles/_sidebar.html.erb
@@ -5,8 +5,8 @@
       <%= render partial: "layouts/main_nav", locals: { context: "sidebar" } %>
       <%= render "layouts/sidebar_tags" %>
     <% end %>
-    <div class="js-display-ad-container" data-async-url="<%= billboard_path(placement_area: :sidebar_left) %>"></div>
-    <div class="js-display-ad-container pt-4" data-async-url="<%= billboard_path(placement_area: :sidebar_left_2) %>"></div>
+    <div class="js-billboard-container" data-async-url="<%= billboard_path(placement_area: :sidebar_left) %>"></div>
+    <div class="js-billboard-container pt-4" data-async-url="<%= billboard_path(placement_area: :sidebar_left_2) %>"></div>
     <% cache("footer-modal-#{ForemInstance.deployed_at}-#{Settings::General.admin_action_taken_at&.rfc3339}", expires_in: 24.hours) do %>
       <%= render "layouts/sidebar_footer" %>
     <% end %>
