+++ app/views/articles/_sticky_nav.html.erb
@@ -55,4 +55,4 @@
   <% end %>
 </div>
 
-<div class="crayons-article-sticky grid gap-4 break-word js-display-ad-container" data-async-url="<%= article_billboard_path(username: @article.username, slug: @article.slug, placement_area: :post_sidebar) %>"></div>
+<div class="crayons-article-sticky grid gap-4 break-word js-billboard-container" data-async-url="<%= article_billboard_path(username: @article.username, slug: @article.slug, placement_area: :post_sidebar) %>"></div>
