+++ app/views/shared/_billboard.html.erb
@@ -1,5 +1,5 @@
 <% if billboard.placement_area.start_with?("feed_") %>
-  <div class="crayons-story crayons-story__display-ad billboard"
+  <div class="crayons-story crayons-story__billboard billboard"
     data-display-unit data-id="<%= billboard.id %>"
     data-category-click="<%= BillboardEvent::CATEGORY_CLICK %>"
     data-category-impression="<%= BillboardEvent::CATEGORY_IMPRESSION %>"
@@ -9,22 +9,22 @@
       <div class="crayons-story__top">
         <%= render partial: "shared/billboard_header", locals: { billboard: billboard } %>
       </div>
-      <div class="crayons-story__indention-display-ad">
-        <div class="text-styles text-styles--display-ad">
+      <div class="crayons-story__indention-billboard">
+        <div class="text-styles text-styles--billboard">
           <%= billboard.processed_html.html_safe %>
         </div>
       </div>
     </div>
   </div>
 <% elsif billboard.placement_area == "home_hero" %>
-  <div class="crayons-card crayons-card--secondary crayons-sponsorship crayons-story__display-ad hero-billboard"
+  <div class="crayons-card crayons-card--secondary crayons-sponsorship crayons-story__billboard hero-billboard"
     data-display-unit data-id="<%= billboard.id %>"
     data-category-click="<%= BillboardEvent::CATEGORY_CLICK %>"
     data-category-impression="<%= BillboardEvent::CATEGORY_IMPRESSION %>"
     data-context-type="<%= data_context_type %>"
     data-type-of="<%= billboard.type_of %>">
     <%= render partial: "shared/billboard_header", locals: { billboard: billboard } %>
-    <div class="p-1 text-styles text-styles--display-ad">
+    <div class="p-1 text-styles text-styles--billboard">
       <%= billboard.processed_html.html_safe %>
     </div>
   </div>
@@ -36,7 +36,7 @@
     data-context-type="<%= data_context_type %>"
     data-type-of="<%= billboard.type_of %>">
     <%= render partial: "shared/billboard_header", locals: { billboard: billboard } %>
-    <div class="p-1 pt-3 text-styles text-styles--display-ad">
+    <div class="p-1 pt-3 text-styles text-styles--billboard">
       <%= billboard.processed_html.html_safe %>
     </div>
   </div>
